pyproject.toml
[tool.poetry]
name = "adilo"
version = "0.1.0"
description = ""
authors = ["FestusOye <festusoyebami@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.1"
djangorestframework = "^3.13.1"
django-status-signal = "^0.0.2"
black = "^22.6.0"
isort = "^5.10.1"
flake8 = "^5.0.4"
virtualenv = "^20.16.3"
django-environ = "^0.9.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
